requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
psycopg2-binary
python-multipart
tzdata